Separated samples and loops from psonic.py
