+++ psonic/api.py
@@ -1,1 +1,21 @@
 from .psonic import *
+from .samples import *
+from .samples.loops import *
+from .samples.ambient import *
+from .samples.bass import *
+from .samples.drums import *
+from .samples.electric import *
+from .samples.guitars import *
+from .samples.misc import *
+from .samples.percussions import *
+from .samples.tabla import *
+from .samples.vinyl import *
+
+
+
+
+
+
+
+
+
