+++ psonic/samples/__init__.py
@@ -1,0 +1,9 @@
+
+class Sample(object):
+    """
+    Sample
+    """
+
+    def __init__(self, name, duration):
+        self.name = name
+        self.duration = duration
