+++ psonic/samples/ambient.py
@@ -1,0 +1,29 @@
+from . import Sample
+
+## Ambient Sounds
+AMBI_CHOIR = Sample('ambi_choir', 1.5715419501133787)
+AMBI_DARK_WOOSH = Sample('ambi_dark_woosh', 3.7021315192743764)
+AMBI_DRONE = Sample('ambi_drone', 4.40843537414966)
+AMBI_GLASS_HUM = Sample('ambi_glass_hum', 10.0)
+AMBI_GLASS_RUB = Sample('ambi_glass_rub', 3.1493650793650794)
+AMBI_HAUNTED_HUM = Sample('ambi_haunted_hum', 9.78156462585034)
+AMBI_LUNAR_LAND = Sample('ambi_lunar_land', 7.394240362811791)
+AMBI_PIANO = Sample('ambi_piano', 2.811746031746032)
+AMBI_SOFT_BUZZ = Sample('ambi_soft_buzz', 0.7821541950113379)
+AMBI_SWOOSH = Sample('ambi_swoosh', 1.8484580498866212)
+
+def get_sounds():
+
+    sounds = dict(
+        AMBI_CHOIR = AMBI_CHOIR,
+        AMBI_DARK_WOOSH = AMBI_DARK_WOOSH,
+        AMBI_DRONE = AMBI_DRONE,
+        AMBI_GLASS_HUM = AMBI_GLASS_HUM,
+        AMBI_GLASS_RUB = AMBI_GLASS_RUB,
+        AMBI_HAUNTED_HUM = AMBI_HAUNTED_HUM,
+        AMBI_LUNAR_LAND = AMBI_LUNAR_LAND,
+        AMBI_PIANO = AMBI_PIANO,
+        AMBI_SOFT_BUZZ = AMBI_SOFT_BUZZ,
+        AMBI_SWOOSH = AMBI_SWOOSH,
+    )
+    return sounds
