+++ psonic/samples/drums.py
@@ -1,0 +1,84 @@
+from . import Sample
+
+## Drum Sounds
+DRUM_BASS_HARD = Sample('drum_bass_hard', 0.6951927437641723)
+DRUM_BASS_SOFT = Sample('drum_bass_soft', 0.5624489795918367)
+DRUM_COWBELL = Sample('drum_cowbell', 0.34988662131519277)
+DRUM_CYMBAL_CLOSED = Sample('drum_cymbal_closed', 0.2069387755102041)
+DRUM_CYMBAL_HARD = Sample('drum_cymbal_hard', 1.6388208616780044)
+DRUM_CYMBAL_OPEN = Sample('drum_cymbal_open', 1.8043764172335601)
+DRUM_CYMBAL_PEDAL = Sample('drum_cymbal_pedal', 0.2721995464852608)
+DRUM_CYMBAL_SOFT = Sample('drum_cymbal_soft', 0.9497278911564626)
+DRUM_HEAVY_KICK = Sample('drum_heavy_kick', 0.2701360544217687)
+DRUM_ROLL = Sample('drum_roll', 6.241678004535148)
+DRUM_SNARE_HARD = Sample('drum_snare_hard', 0.44492063492063494)
+DRUM_SNARE_SOFT = Sample('drum_snare_soft', 0.3147845804988662)
+DRUM_SPLASH_HARD = Sample('drum_splash_hard', 2.5961904761904764)
+DRUM_SPLASH_SOFT = Sample('drum_splash_soft', 1.857981859410431)
+DRUM_TOM_HI_HARD = Sample('drum_tom_hi_hard', 0.7376643990929705)
+DRUM_TOM_HI_SOFT = Sample('drum_tom_hi_soft', 0.6290249433106576)
+DRUM_TOM_LO_HARD = Sample('drum_tom_lo_hard', 1.0002267573696144)
+DRUM_TOM_LO_SOFT = Sample('drum_tom_lo_soft', 0.8634013605442177)
+DRUM_TOM_MID_HARD = Sample('drum_tom_mid_hard', 0.7342176870748299)
+DRUM_TOM_MID_SOFT = Sample('drum_tom_mid_soft', 0.6721315192743764)
+
+## Snare Drums Sounds
+SN_DOLF = Sample('sn_dolf', 0.37759637188208617)
+SN_DUB = Sample('sn_dub', 0.2781179138321995)
+SN_ZOME = Sample('sn_zome', 0.4787528344671202)
+
+## Bass Drums Sounds
+BD_808 = Sample('bd_808', 0.5597505668934241)
+BD_ADA = Sample('bd_ada', 0.10179138321995465)
+BD_BOOM = Sample('bd_boom', 1.7142857142857142)
+BD_FAT = Sample('bd_fat', 0.23219954648526078)
+BD_GAS = Sample('bd_gas', 0.4471428571428571)
+BD_HAUS = Sample('bd_haus', 0.21993197278911564)
+BD_KLUB = Sample('bd_klub', 0.368843537414966)
+BD_PURE = Sample('bd_pure', 0.43324263038548755)
+BD_SONE = Sample('bd_sone', 0.4089115646258503)
+BD_TEK = Sample('bd_tek', 0.24024943310657595)
+BD_ZOME = Sample('bd_zome', 0.45972789115646256)
+BD_ZUM = Sample('bd_zum', 0.13158730158730159)
+
+def get_sounds():
+    sounds = dict(
+        DRUM_BASS_HARD = DRUM_BASS_HARD,
+        DRUM_BASS_SOFT = DRUM_BASS_SOFT,
+        DRUM_COWBELL = DRUM_COWBELL,
+        DRUM_CYMBAL_CLOSED = DRUM_CYMBAL_CLOSED,
+        DRUM_CYMBAL_HARD = DRUM_CYMBAL_HARD,
+        DRUM_CYMBAL_OPEN = DRUM_CYMBAL_OPEN,
+        DRUM_CYMBAL_PEDAL = DRUM_CYMBAL_PEDAL,
+        DRUM_CYMBAL_SOFT = DRUM_CYMBAL_SOFT,
+        DRUM_HEAVY_KICK = DRUM_HEAVY_KICK,
+        DRUM_ROLL = DRUM_ROLL,
+        DRUM_SNARE_HARD = DRUM_SNARE_HARD,
+        DRUM_SNARE_SOFT = DRUM_SNARE_SOFT,
+        DRUM_SPLASH_HARD = DRUM_SPLASH_HARD,
+        DRUM_SPLASH_SOFT = DRUM_SPLASH_SOFT,
+        DRUM_TOM_HI_HARD = DRUM_TOM_HI_HARD,
+        DRUM_TOM_HI_SOFT = DRUM_TOM_HI_SOFT,
+        DRUM_TOM_LO_HARD = DRUM_TOM_LO_HARD,
+        DRUM_TOM_LO_SOFT = DRUM_TOM_LO_SOFT,
+        DRUM_TOM_MID_HARD = DRUM_TOM_MID_HARD,
+        DRUM_TOM_MID_SOFT = DRUM_TOM_MID_SOFT,
+
+        SN_DOLF = SN_DOLF,
+        SN_DUB = SN_DUB,
+        SN_ZOME = SN_ZOME,
+
+        BD_808 = BD_808,
+        BD_ADA = BD_ADA,
+        BD_BOOM = BD_BOOM,
+        BD_FAT = BD_FAT,
+        BD_GAS = BD_GAS,
+        BD_HAUS = BD_HAUS,
+        BD_KLUB = BD_KLUB,
+        BD_PURE = BD_PURE,
+        BD_SONE = BD_SONE,
+        BD_TEK = BD_TEK,
+        BD_ZOME = BD_ZOME,
+        BD_ZUM = BD_ZUM,
+    )
+    return sounds
