+++ psonic/samples/electric.py
@@ -1,0 +1,58 @@
+from . import Sample
+
+## Electric Sounds
+ELEC_BEEP = Sample('elec_beep', 0.10578231292517007)
+ELEC_BELL = Sample('elec_bell', 0.27825396825396825)
+ELEC_BLIP = Sample('elec_blip', 0.15816326530612246)
+ELEC_BLIP2 = Sample('elec_blip2', 0.1840816326530612)
+ELEC_BLUP = Sample('elec_blup', 0.6632199546485261)
+ELEC_BONG = Sample('elec_bong', 0.3464852607709751)
+ELEC_CHIME = Sample('elec_chime', 2.2775510204081635)
+ELEC_CYMBAL = Sample('elec_cymbal', 0.563219954648526)
+ELEC_FILT_SNARE = Sample('elec_filt_snare', 1.5158503401360544)
+ELEC_FLIP = Sample('elec_flip', 0.07476190476190477)
+ELEC_FUZZ_TOM = Sample('elec_fuzz_tom', 0.37179138321995464)
+ELEC_HI_SNARE = Sample('elec_hi_snare', 0.19736961451247165)
+ELEC_HOLLOW_KICK = Sample('elec_hollow_kick', 0.15564625850340136)
+ELEC_LO_SNARE = Sample('elec_lo_snare', 0.6607936507936508)
+ELEC_MID_SNARE = Sample('elec_mid_snare', 0.7256235827664399)
+ELEC_PING = Sample('elec_ping', 0.21226757369614513)
+ELEC_PLIP = Sample('elec_plip', 0.19882086167800453)
+ELEC_POP = Sample('elec_pop', 0.08680272108843537)
+ELEC_SNARE = Sample('elec_snare', 0.3893197278911565)
+ELEC_SOFT_KICK = Sample('elec_soft_kick', 0.1364172335600907)
+ELEC_TICK = Sample('elec_tick', 0.01943310657596372)
+ELEC_TRIANGLE = Sample('elec_triangle', 0.22294784580498866)
+ELEC_TWANG = Sample('elec_twang', 0.6243083900226757)
+ELEC_TWIP = Sample('elec_twip', 0.10140589569160997)
+ELEC_WOOD = Sample('elec_wood', 0.47811791383219954)
+
+def get_sounds():
+    sounds = dict(
+        ELEC_BEEP = ELEC_BEEP,
+        ELEC_BELL = ELEC_BELL,
+        ELEC_BLIP = ELEC_BLIP,
+        ELEC_BLIP2 = ELEC_BLIP2,
+        ELEC_BLUP = ELEC_BLUP,
+        ELEC_BONG = ELEC_BONG,
+        ELEC_CHIME = ELEC_CHIME,
+        ELEC_CYMBAL = ELEC_CYMBAL,
+        ELEC_FILT_SNARE = ELEC_FILT_SNARE,
+        ELEC_FLIP = ELEC_FLIP,
+        ELEC_FUZZ_TOM = ELEC_FUZZ_TOM,
+        ELEC_HI_SNARE = ELEC_HI_SNARE,
+        ELEC_HOLLOW_KICK = ELEC_HOLLOW_KICK,
+        ELEC_LO_SNARE = ELEC_LO_SNARE,
+        ELEC_MID_SNARE = ELEC_MID_SNARE,
+        ELEC_PING = ELEC_PING,
+        ELEC_PLIP = ELEC_PLIP,
+        ELEC_POP = ELEC_POP,
+        ELEC_SNARE = ELEC_SNARE,
+        ELEC_SOFT_KICK = ELEC_SOFT_KICK,
+        ELEC_TICK = ELEC_TICK,
+        ELEC_TRIANGLE = ELEC_TRIANGLE,
+        ELEC_TWANG = ELEC_TWANG,
+        ELEC_TWIP = ELEC_TWIP,
+        ELEC_WOOD = ELEC_WOOD,
+    )
+    return sounds
