+++ psonic/samples/guitars.py
@@ -1,0 +1,18 @@
+from . import Sample
+
+## Sounds featuring guitars
+GUIT_E_FIFTHS = Sample('guit_e_fifths', 5.971791383219955)
+GUIT_E_SLIDE = Sample('guit_e_slide', 4.325192743764172)
+GUIT_EM9 = Sample('guit_em9', 9.972063492063493)
+GUIT_HARMONICS = Sample('guit_harmonics', 3.5322675736961453)
+
+def get_sounds():
+
+    sounds = dict(
+        GUIT_E_FIFTHS = GUIT_E_FIFTHS,
+        GUIT_E_SLIDE = GUIT_E_SLIDE,
+        GUIT_EM9 = GUIT_EM9,
+        GUIT_HARMONICS = GUIT_HARMONICS,
+    )
+    return sounds
+
