+++ psonic/samples/loops/__init__.py
@@ -1,0 +1,12 @@
+from psonic.samples import Sample
+
+## Sounds for Looping
+LOOP_AMEN = Sample('loop_amen', 1.753310657596372)
+LOOP_AMEN_FULL = Sample('loop_amen_full', 6.857142857142857)
+LOOP_BREAKBEAT = Sample('loop_breakbeat', 1.9047619047619047)
+LOOP_COMPUS = Sample('loop_compus', 6.486485260770975)
+LOOP_GARZUL = Sample('loop_garzul', 8.0)
+LOOP_INDUSTRIAL = Sample('loop_industrial', 0.8837414965986394)
+LOOP_MIKA = Sample('loop_mika', 8.0)
+LOOP_SAFARI = Sample('loop_safari', 8.005079365079364)
+LOOP_TABLA = Sample('loop_tabla', 10.673990929705216)
