+++ psonic/samples/tabla.py
@@ -1,0 +1,62 @@
+from . import Sample
+
+## Tabla
+TABLA_DHEC = Sample('tabla_dhec', 0.3250793650793651)
+TABLA_GHE1 = Sample('tabla_ghe1', 0.5912244897959184)
+TABLA_GHE2 = Sample('tabla_ghe2', 2.6607256235827665)
+TABLA_GHE3 = Sample('tabla_ghe3', 2.3908163265306124)
+TABLA_GHE4 = Sample('tabla_ghe4', 0.7960997732426304)
+TABLA_GHE5 = Sample('tabla_ghe5', 3.560045351473923)
+TABLA_GHE6 = Sample('tabla_ghe6', 3.6011337868480724)
+TABLA_GHE7 = Sample('tabla_ghe7', 2.1512698412698414)
+TABLA_GHE8 = Sample('tabla_ghe8', 0.5817913832199546)
+TABLA_KE1 = Sample('tabla_ke1', 0.04342403628117914)
+TABLA_KE2 = Sample('tabla_ke2', 0.04403628117913832)
+TABLA_KE3 = Sample('tabla_ke3', 0.07571428571428572)
+TABLA_NA = Sample('tabla_na', 0.7198185941043084)
+TABLA_NA_O = Sample('tabla_na_o', 1.4889795918367348)
+TABLA_NA_S = Sample('tabla_na_s', 0.23582766439909297)
+TABLA_RE = Sample('tabla_re', 0.2815419501133787)
+TABLA_TAS1 = Sample('tabla_tas1', 1.1116553287981858)
+TABLA_TAS2 = Sample('tabla_tas2', 1.4338321995464853)
+TABLA_TAS3 = Sample('tabla_tas3', 1.2364625850340136)
+TABLA_TE1 = Sample('tabla_te1', 0.17777777777777778)
+TABLA_TE2 = Sample('tabla_te2', 0.33233560090702946)
+TABLA_TE_M = Sample('tabla_te_m', 0.28879818594104306)
+TABLA_TE_NE = Sample('tabla_te_ne', 0.15310657596371882)
+TABLA_TUN1 = Sample('tabla_tun1', 2.3394104308390022)
+TABLA_TUN2 = Sample('tabla_tun2', 2.693514739229025)
+TABLA_TUN3 = Sample('tabla_tun3', 2.0956009070294783)
+
+def get_sounds():
+
+    sounds = dict(
+        TABLA_DHEC = TABLA_DHEC,
+        TABLA_GHE1 = TABLA_GHE1,
+        TABLA_GHE2 = TABLA_GHE2,
+        TABLA_GHE3 = TABLA_GHE3,
+        TABLA_GHE4 = TABLA_GHE4,
+        TABLA_GHE5 = TABLA_GHE5,
+        TABLA_GHE6 = TABLA_GHE6,
+        TABLA_GHE7 = TABLA_GHE7,
+        TABLA_GHE8 = TABLA_GHE8,
+        TABLA_KE1 = TABLA_KE1,
+        TABLA_KE2 = TABLA_KE2,
+        TABLA_KE3 = TABLA_KE3,
+        TABLA_NA = TABLA_NA,
+        TABLA_NA_O = TABLA_NA_O,
+        TABLA_NA_S = TABLA_NA_S,
+        TABLA_RE = TABLA_RE,
+        TABLA_TAS1 = TABLA_TAS1,
+        TABLA_TAS2 = TABLA_TAS2,
+        TABLA_TAS3 = TABLA_TAS3,
+        TABLA_TE1 = TABLA_TE1,
+        TABLA_TE2 = TABLA_TE2,
+        TABLA_TE_M = TABLA_TE_M,
+        TABLA_TE_NE = TABLA_TE_NE,
+        TABLA_TUN1 = TABLA_TUN1,
+        TABLA_TUN2 = TABLA_TUN2,
+        TABLA_TUN3 = TABLA_TUN3,
+    )
+    return sounds
+
